Refactored Lexer.scan().
Illegal characters are not ignored anymore. They are reported as errors.

Added a new member 'ws' to Token. When a token is scanned the lexer sets
ws to the leading whitespace or leaves it at null when no whitespace was found.

Added Illegal to enum TOK and IllegalCharacter to enum MID.
Added localized messages for MID.IllegalCharacter.

Adapted code of cmd.Generate to make use of Token.ws.