Some fixes and changes to doc gen files.
Added a title tag to the HTML text in html_map.d.
Added SYMLINK macro. Added attributes beg and end to the link in SYMBOL.
Calls to Path.rmtree() are with no errors by default.
Remove .svn directories copied from an SVN checkout of Tango.
Give syntax highlighted files the name of the file as the title.