DDoc: use an internal format for macros in a text.
* Instead of emitting, e.g., "$(MacroName ...)", emit "\1MacroName ...\2".
These special marker characters indicate where macros begin and end.
('\1' and '\2' are octal escape sequences; their int values are 1 and 2.)
As a result, unbalanced parentheses or unclosed macros in Ddoc comments
do not affect the rest of the macros in the text.
Less meaningful warnings are avoided by this.
Another advantage is that '(' and ')' do not have to be replaced
with $(LP) and $(RP) respectively.

* Added function String(char*, char*) to dil.lexer.Funcs.
  It replaces makeString().